pybind/mgr: move build_url from dashboard to mgr_util
(cherry picked from commit 29883d68ef696a361b4fec0af22505835170a74d)
